api/requirements.txt
fastapi==0.135.3
uvicorn==0.41.0
google-generativeai==0.8.6
PyMuPDF==1.27.2
numpy==2.4.4
python-dotenv==1.2.2
python-multipart==0.0.22
pymongo==4.9.0
mangum==0.17.0